requirement.txt
aspy.yaml==1.1.2
awscli==1.16.105
awsebcli==3.14.11
botocore==1.12.95
cement==2.8.2
certifi==2018.11.29
cfgv==1.4.0
chardet==3.0.4
Click==7.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-rest-swagger==2.2.0
djangorestframework==3.9.4
docutils==0.14
filelock==3.0.10
identify==1.3.0
idna==2.7
importlib-metadata==0.8
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
lxml==4.3.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
nodeenv==1.3.3
openapi-codec==1.3.2
pathspec==0.5.5
pluggy==0.9.0
pre-commit==1.14.4
protobuf==3.8.0
psycopg2==2.8.3
py==1.8.0
pyasn1==0.4.5
pycryptodome==3.8.1
pymemcache==2.1.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.20.1
requests-toolbelt==0.9.1
rsa==3.4.2
s3transfer==0.2.0
semantic-version==2.5.0
simplejson==3.16.0
six==1.11.0
South==1.0.2
termcolor==1.1.0
toml==0.10.0
tox==3.7.0
translate==3.5.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.4.1
xlrd==1.2.0
zipp==0.3.3
